Fix timeline editing bug and update UI
Refactored the timeline editing logic to fix a bug where timelineEditing wasn't properly set to true when the user began touching the slider. Also updated the UI to display the current playback time and to only update the timeline value and current time label when timelineEditing is false. Finally, added logic to remove all observers when the view controller disappears.