Add recommendation queueing functionality to PlayerViewController
This commit adds functionality to queue up recommended songs after the currently playing song in the PlayerViewController. When a song is played for the first time, a call to the queueRecommendation method is made to fetch a list of recommended songs based on the currently playing song. The first recommended song is then inserted into the player's queue using the insert method.

The queueRecommendation method uses the MusadoraKit library to fetch recommended songs for the currently playing song. The fetched song is then inserted into the player's queue, and will be played once the current song has finished.

This new functionality adds to the user experience of the app, as it allows users to easily discover new music based on what they are currently listening to.