Improved repeat mode toggle animation
Updated the toggleRepeatMode() function to add a smooth transition between the different repeat mode icons, with a slight bounce effect. Additionally, added a reusable function transitionToPurple() that transitions the color of an image from white to a custom UIColor named "Purple". This function is used to update the color of the repeat mode icons when the repeat mode is set to "all" or "one".